chore: move files to folders
