+++ meteor/server/api/rest/v0/index.ts
@@ -2,13 +2,13 @@ import { ServerResponse, IncomingMessage } from 'http'
 import { Params } from 'meteor/meteorhacks:picker'
 import * as _ from 'underscore'
 import { Meteor } from 'meteor/meteor'
-import { MeteorMethodSignatures } from '../../methods'
-import { PubSub } from '../../../lib/api/pubsub'
-import { MeteorPublications, MeteorPublicationSignatures } from '../../publications/lib'
-import { UserActionAPIMethods } from '../../../lib/api/userActions'
-import { PickerPOST, PickerGET, AsyncRouter } from '../http'
-import { logger } from '../../../lib/logging'
-import { ClientAPI } from '../../../lib/api/client'
+import { MeteorMethodSignatures } from '../../../methods'
+import { PubSub } from '../../../../lib/api/pubsub'
+import { MeteorPublications, MeteorPublicationSignatures } from '../../../publications/lib'
+import { UserActionAPIMethods } from '../../../../lib/api/userActions'
+import { PickerPOST, PickerGET, AsyncRouter } from '../../http'
+import { logger } from '../../../../lib/logging'
+import { ClientAPI } from '../../../../lib/api/client'
 
 const apiVersion = 0
